Add new and old order callback classes
